test(rbac): comprehensive manifest-driven RBAC matrix + full-stack e2e
The RBAC coverage had two gaps: the comprehensive check (authorization-matrix
.spec.ts) was static (derives verdicts from decorators + policies, never hits a
server), while the only live e2e check covered just 6 hand-picked endpoints —
and skipped tier-gated features (Cockpit/SSO) entirely because the e2e org was
free-tier, so their RBAC was never actually exercised.

Close both with a single source of truth:

- rbac-matrix.shared.ts: extract every gated endpoint's @CheckPolicies (TS
  compiler) AND its HTTP method + URL path, then evaluate the verdict per role
  with the REAL PermissionsAbilityFactory + handlers. Both consumers below run
  off this one extractor, so the static grid and the live e2e can never
  disagree about what the matrix says.
- authorization-matrix.spec.ts: refactored to consume the shared extractor
  (drops the duplicated ~360 lines); same assertions.
- rbac-matrix.manifest.json + rbac-matrix.manifest.spec.ts: a committed
  manifest (every endpoint × role → allow/deny) with a jest drift-guard that
  fails CI if a controller's gating changes without regenerating it
  (UPDATE_RBAC_MANIFEST=1).
- tests/e2e/scenarios/rbac-authorization.ts: rewritten to REPLAY the full
  manifest against a running API (was 6 endpoints, now all 142). Runs on a
  trial (fresh cloud org) / licensed target so the tier-gated guards are
  unlocked and their RBAC is actually tested. OWNER is the canary: an owner
  401/403 can only be a non-RBAC guard, so those endpoints are reported and
  skipped (never silently passed); if over half are owner-blocked the run
  fails loudly (org not tier-unlocked → tier-gated RBAC not validated).

Refs #1229.
